+++ Risicammara-java/src/risicammaraServer/messaggiManage/MessaggioAddPlayer.java
@@ -33,7 +33,7 @@ public class MessaggioAddPlayer implements Messaggio{
 
 
     @Override
-    public int getSender() {
+    public long getSender() {
         return -1;
     }
