+++ Risicammara-java/src/risicammaraServer/messaggiManage/MessaggioCarta.java
@@ -9,14 +9,14 @@ import risicammaraJava.deckManage.Carta;
  */
 public class MessaggioCarta implements Messaggio {
     private Carta carta;
-    private int sender;
+    private long sender;
 
     /**
      * Crea un messaggio Carta
      * @param carta la carta da inviare
      * @param sender chi ha preso la carta.
      */
-    public MessaggioCarta(Carta carta,int sender) {
+    public MessaggioCarta(Carta carta,long sender) {
         this.carta = carta;
         this.sender = sender;
     }
@@ -35,7 +35,7 @@ public class MessaggioCarta implements Messaggio {
     }
 
     @Override
-    public int getSender() {
+    public long getSender() {
         return sender;
     }
